Rename `tests/emit` to `tests/emit_json'.
