+++ tests/emit_json.rs
