+++ tests/lib.rs
@@ -6,7 +6,7 @@ extern crate std;
 
 pub mod deserialize;
 pub mod deserialize_sets;
-pub mod emit;
+pub mod emit_json;
 pub mod parse;
 pub mod parse_sets;
 pub mod ser;
